Master updated

+++ src/test/java/io/confluent/connect/elasticsearch/integration/ElasticsearchConnectorBaseIT.java
@@ -104,7 +104,7 @@ public class ElasticsearchConnectorBaseIT extends BaseConnectorIT {
   public void cleanup() throws IOException {
     stopConnect();
 
-    if (container.isRunning()) {
+    if (container != null && container.isRunning()) {
       if (helperClient != null) {
         try {
           helperClient.deleteIndex(index, isDataStream);
@@ -178,7 +178,8 @@ public class ElasticsearchConnectorBaseIT extends BaseConnectorIT {
     props.put(CONNECTION_URL_CONFIG, address);
     helperClient = new ElasticsearchHelperClient(
         props.get(CONNECTION_URL_CONFIG),
-        new ElasticsearchSinkConnectorConfig(props)
+        new ElasticsearchSinkConnectorConfig(props),
+        container.shouldStartClientInCompatibilityMode()
     );
   }
 
@@ -250,13 +251,14 @@ public class ElasticsearchConnectorBaseIT extends BaseConnectorIT {
         .indices("*")
         .privileges("create_index", "read", "write", "view_index_metadata")
         .build();
-    Builder builder = Role.builder();
-    builder = forDataStream ? builder.clusterPrivileges("monitor") : builder;
-    Role role = builder
+    // Historically (i.e. ES the previous test base version 7.9.3), ES_SINK_CONNECTOR_ROLE would not require the
+    // "monitor" cluster privilege.  However, this has changed for 7.16.3, although leaving the surrounding
+    // logic in place in the case that future ES versions or tests wish to diverge the permissions.
+    return Role.builder()
         .name(forDataStream ? ES_SINK_CONNECTOR_DS_ROLE : ES_SINK_CONNECTOR_ROLE)
         .indicesPrivileges(indicesPrivileges)
+        .clusterPrivileges("monitor")
         .build();
-    return role;
   }
 
   private static User getMinimalPrivilegesUser(boolean forDataStream) {
